requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
cryptography==2.9.2
Django==2.0
jwcrypto==0.7
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
